restructuring the classes
